Implements antctl query endpoint
Antctl query endpoint gives users the ability to filter
network policies relevant to a certain endpoint. The
user is able to query internal controller stores
using the antrea controller apiserver through
antctl.

To answer queries in an efficient way, we add new indexers
to the internal stores of the controller, incurring a
small memory overhead (~20%, depends on structure of
NetworkPolicies).

Note, antctl query endpoint only works when run in the controller.
We are working on getting it to work out of cluster.

Fixes #974